requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.10.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.1.4
django-ckeditor==6.5.1
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
exceptiongroup==1.1.3
filelock==3.13.1
flake8==6.1.0
fonttools==4.45.0
fpdf2==2.6.0
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
IP2Location==8.8.0
isort==5.13.2
kaitaistruct==0.10
mccabe==0.7.0
multidict==6.0.4
mypy==1.9.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pika==1.3.2
Pillow==9.3.0
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pyOpenSSL==22.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rucaptcha==3.0.1
redis==4.3.4
requests==2.31.0
requests-file==1.5.1
selenium==4.18.1
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==3.3.1
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
types-requests==2.31.0.20240311
typing_extensions==4.10.0
urllib3==2.2.1
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.3
zstandard==0.22.0
